backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.4
python-jose==3.3.0
passlib==1.7.4
slowapi==0.1.9
duckduckgo-search==5.3.1
google-genai==0.3.0